Implement WSL setup script with zsh and docker integration
