+++ go.mod
@@ -3,6 +3,23 @@ module github.com/kachick/dotfiles
 go 1.20
 
 require (
+	github.com/foxboron/sbctl v0.0.0-20230803204657-ef78d240a5e9
 	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b
 	golang.org/x/sys v0.11.0
+)
+
+require (
+	github.com/fatih/color v1.13.0 // indirect
+	github.com/foxboron/go-uefi v0.0.0-20230217160721-375279dfc837 // indirect
+	github.com/google/certificate-transparency-go v1.1.2 // indirect
+	github.com/google/go-attestation v0.3.3-0.20210901195502-505680f536da // indirect
+	github.com/google/go-tpm v0.3.2 // indirect
+	github.com/google/go-tspi v0.3.0 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/spf13/afero v1.9.3 // indirect
+	golang.org/x/crypto v0.6.0 // indirect
+	golang.org/x/text v0.7.0 // indirect
 )
